requirements/requirements-base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/requirements-base.in
#
-e git+https://github.com/SALTISES4/CourseFlow#egg=CourseFlow  # via -r requirements/requirements-base.in
-e git+https://github.com/aahilgert/UserFeedback@master#egg=user_feedback  # via -r requirements/requirements-base.in
-e git+https://github.com/SALTISES4/django-cookies-samesite#egg=django_cookies_samesite  # via -r requirements/requirements-base.in
-e git+https://github.com/SALTISES4/django-lti-tool-provider@dev#egg=django_lti_tool_provider  # via -r requirements/requirements-base.in
-e git+https://github.com/edx/opaque-keys#egg=opaque-keys  # via -r requirements/requirements-base.in
-e git+https://github.com/SALTISES4/pinax-forums@dev#egg=pinax_forums  # via -r requirements/requirements-base.in
amqp==2.5.2               # via kombu
beautifulsoup4==4.8.2     # via django-htmlmin
billiard==3.6.3.0         # via celery
bleach==3.1.4             # via -r requirements/requirements-base.in
celery[redis]==4.4.0      # via -r requirements/requirements-base.in, django-celery-beat
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
django-analytical==2.5.0  # via -r requirements/requirements-base.in
django-appconf==1.0.3     # via django-compressor, pinax-forums
django-celery-beat==2.0.0  # via -r requirements/requirements-base.in
django-classy-tags==1.0.0  # via django-cookie-law
django-compressor==2.4    # via -r requirements/requirements-base.in
django-cookie-law==2.0.3  # via -r requirements/requirements-base.in
django-cookies-samesite-fork==0.2.1  # via django-security-headers
django-csp==3.6           # via django-security-headers
django-grappelli==2.14.1  # via -r requirements/requirements-base.in
django-htmlmin==0.11.0    # via -r requirements/requirements-base.in
django-security-headers==0.3.1  # via -r requirements/requirements-base.in
django-sslserver==0.22    # via django-security-headers
django-timezone-field==4.0  # via django-celery-beat
django==2.2.11            # via -r requirements/requirements-base.in, django-analytical, django-appconf, django-celery-beat, django-classy-tags, django-cookie-law, django-csp, django-lti-tool-provider, django-security-headers, django-sslserver, django-timezone-field, djangorestframework, jsonfield, pinax-forums
djangorestframework==3.11.0  # via -r requirements/requirements-base.in
filetype==1.0.5           # via pinax-forums
html5lib==1.0.1           # via django-htmlmin
idna==2.9                 # via requests
importlib-metadata==1.5.0  # via kombu
jsonfield==3.1.0          # via django-lti-tool-provider
kombu==4.6.7              # via celery
lti==0.9.4                # via django-lti-tool-provider
lxml==4.5.0               # via lti
mysqlclient==1.4.6        # via -r requirements/requirements-base.in
numpy==1.18.1             # via pandas
oauthlib==3.1.0           # via django-lti-tool-provider, lti, requests-oauthlib
packaging==20.1           # via -r requirements/requirements-base.in
pandas==1.0.2             # via -r requirements/requirements-base.in
pbr==5.4.4                # via stevedore
pillow==7.0.0             # via -r requirements/requirements-base.in
pyjwt==1.7.1              # via -r requirements/requirements-base.in
pymongo==3.10.1           # via edx-opaque-keys
pyparsing==2.4.6          # via packaging
pytest-runner==5.2        # via django-cookie-law
python-crontab==2.4.0     # via django-celery-beat
python-dateutil==2.8.1    # via pandas, python-crontab
python-memcached==1.59    # via -r requirements/requirements-base.in
pytz==2019.3              # via -r requirements/requirements-base.in, celery, django, django-timezone-field, pandas
rcssmin==1.0.6            # via django-compressor
redis==3.4.1              # via celery
requests-oauthlib==1.3.0  # via lti
requests==2.23.0          # via -r requirements/requirements-base.in, requests-oauthlib
rjsmin==1.1.0             # via django-compressor
six==1.14.0               # via bleach, django-appconf, django-classy-tags, django-compressor, django-grappelli, django-lti-tool-provider, edx-opaque-keys, html5lib, packaging, python-dateutil, python-memcached, stevedore
soupsieve==2.0            # via beautifulsoup4
sqlparse==0.3.1           # via django
stevedore==1.32.0         # via edx-opaque-keys
urllib3==1.25.8           # via requests
vine==1.3.0               # via amqp, celery
webencodings==0.5.1       # via bleach, html5lib
zipp==3.0.0               # via importlib-metadata
